[CLEANUP] Cleaned up angular artifacts. - cleaned up names of angular modules; - removed empty files; - replaced using default module exports to avoid possible confusing in the future; - fixed karma config to run tests; -combined component  +controller files to single one.
