+++ step-3/client/app/app.component.js
@@ -1,9 +1,6 @@
 import template from './app.html';
 import './app.scss';
 
-let appComponent = {
-  template,
-  restrict: 'E'
+export const AppComponent = {
+  template
 };
-
-export default appComponent;
