+++ step-3/client/app/app.html
@@ -1,5 +1,4 @@
 <div class="app">
-    <navigation></navigation>
-    
-    <div ui-view></div>
+  <navigation></navigation>
+  <div ui-view></div>
 </div>
