+++ step-3/client/app/app.scss
@@ -41,4 +41,3 @@ li.ui-state-default:last-child{
 .marginTop10px{
     margin-top:10px;
 }
-
