+++ step-3/client/app/components/home/home.component.js
@@ -1,12 +1,6 @@
 import template from './home.html';
-import controller from './home.controller';
 import './home.scss';
 
-let homeComponent = {
-  restrict: 'E',
-  bindings: {},
-  template,
-  controller
+export const HomeComponent = {
+  template
 };
-
-export default homeComponent;
