--- step-3/client/app/components/home/home.controller.js
@@ -1,6 +1,0 @@
-class HomeController {
-  constructor() {
-  }
-}
-
-export default HomeController;
