+++ step-3/client/app/components/home/home.html
@@ -1,7 +1,7 @@
 <div class="container">
-    <div class="row">
-        <div class="col-md-12 center">
-            This is the Homepage of our simple To-do app
-        </div>
+  <div class="row">
+    <div class="col-md-12 center">
+      This is the Homepage of our simple To-do app
     </div>
+  </div>
 </div>
