--- step-3/client/app/components/home/home.js
@@ -1,9 +1,0 @@
-import angular from 'angular';
-import uiRouter from '@uirouter/angularjs';
-import homeComponent from './home.component';
-
-const homeModule = angular.module('home', [uiRouter])
-
-.component('home', homeComponent);
-
-export default homeModule;
