+++ step-3/client/app/components/navigation/navigation.component.js
@@ -1,13 +1,5 @@
 import template from './navigation.html';
-import controller from './navigation.controller';
-import './navigation.scss';
 
-let navigationComponent = {
-  restrict: 'E',
-  bindings: {},
-  template,
-  controller,
-  controllerAs: 'vm'
+export const NavigationComponent = {
+  template
 };
-
-export default navigationComponent;
