--- step-3/client/app/components/navigation/navigation.controller.js
@@ -1,7 +1,0 @@
-class NavigationController {
-  constructor() {
-
-  }
-}
-
-export default NavigationController;
