+++ step-3/client/app/components/navigation/navigation.module.js
@@ -1,0 +1,8 @@
+import angular from 'angular';
+import uiRouter from '@uirouter/angularjs';
+import { NavigationComponent } from './navigation.component';
+
+export const NavigationModule = angular
+  .module('navigation', [uiRouter])
+  .component('navigation', NavigationComponent)
+  .name;
