+++ step-3/client/app/components/noredux/noredux.component.js
@@ -1,12 +1,63 @@
 import template from './noredux.html';
-import controller from './noredux.controller';
 import './noredux.scss';
 
-let noreduxComponent = {
-  restrict: 'E',
-  bindings: {},
+export const NoReduxComponent = {
   template,
-  controller
-};
+  controller: class NoReduxController {
+    constructor() {
+      // Properties
+      this.todos = [
+        { text: 'First todo item', done: false },
+        { text: 'Second todo item', done: false }
+      ];
+      this.todo = '';
+      this.showDone = false;
+    }
 
-export default noreduxComponent;
+    // Actions
+    addTodo() {
+      this.todos.push({ text: this.todo, done: false });
+      this.todo = '';
+    }
+
+    deleteTodo(index) {
+      this.todos.splice(index, 1);
+    }
+
+    toggleDone(index) {
+      this.todos[index].done = !this.todos[index].done;
+    }
+
+    getTotalTodos() {
+      return this.todos.length;
+    }
+
+    getTotalDoneTodos() {
+      return this.todos.filter(function (todo) {
+        return todo.done;
+      }).length;
+    }
+
+    getAllDone() {
+      return this.todos.filter(function (todo) {
+        return todo.done;
+      });
+    }
+
+    deleteAllDone() {
+      this.todos = this.todos.filter(function (todo) {
+        return !todo.done;
+      });
+    }
+
+    markAllAsDone() {
+      var isAnyMarkedAsDone = this.todos.filter(function  (todo) {
+        return todo.done;
+      }).length > 0;
+
+      this.todos.map(function (t) {
+        t.done = !isAnyMarkedAsDone;
+      });
+    }
+  }
+};
