+++ step-3/client/app/components/noredux/noredux.html
@@ -1,64 +1,62 @@
-<code>
-    {{$ctrl}}
-</code>
+<code>{{$ctrl}}</code>
 
 <div class="container">
-    <div class="row">
-        <div class="col-md-12 text-center">
-            <h1>Without Redux</h1>
-        </div>
-        <div class="col-md-6">
-            <div class="todolist not-done">
-                <h1>To-do list</h1>
-                <div class="input-group">
-                    <input type="text" class="form-control add-todo" placeholder="Add todo" ng-model="$ctrl.todo">
-                    <span class="input-group-btn">
-                    <button class="btn btn-primary" type="button" ng-disabled="!$ctrl.todo" ng-click="$ctrl.addTodo()">Add</button>
-                  </span>
-                </div>
-
-                <button class="btn btn-success marginTop10px" ng-click="$ctrl.markAllAsDone()">Mark all as done</button>
-                <button class="btn btn-danger marginTop10px" ng-click="$ctrl.deleteAllDone()">Delete all done</button>
-                <div class="checkbox pull-right">
-                    <label>
-                        <input type="checkbox" ng-model="$ctrl.showDone" />
-                        Show done
-                    </label>
-                </div>
-                <hr>
-
-                <ul id="sortable" class="list-unstyled">
-                    <li class="ui-state-default" ng-repeat="todo in $ctrl.todos" ng-if="!(!$ctrl.showDone && todo.done)">
-                        <div class="checkbox">
-                            <label class="done-{{todo.done}}">
-                                <input type="checkbox" ng-model="todo.done" />
-                                {{todo.text}}
-                            </label>
-                            <button class="remove-item btn btn-danger btn-xs pull-right" ng-click="$ctrl.deleteTodo($index)">
-                                <span class="glyphicon glyphicon-remove"></span>
-                            </button>
-                        </div>
-                    </li>
-                </ul>
-                <div class="todo-footer">
-                    <strong><span class="count-todos">{{$ctrl.getTotalTodos()}}</span></strong> total |
-                    <strong><span class="count-todos">{{$ctrl.getTotalDoneTodos()}}</span></strong> done
-                </div>
-            </div>
-        </div>
-        <div class="col-md-6">
-            <div class="todolist">
-                <h1>Already done</h1>
-                <button class="btn btn-warning marginTop10px" ng-click="$ctrl.markAllAsDone(false)">Uncheck all</button>
-                <ul id="done-items" class="list-unstyled">
-                    <li ng-repeat="doneTodo in $ctrl.getAllDone()">
-                        {{doneTodo.text}}
-                        <button class="remove-item btn btn-danger btn-xs pull-right" ng-click="$ctrl.toggleDone($index)">
-                            <span class="glyphicon glyphicon-remove"></span>
-                        </button>
-                    </li>
-                </ul>
-            </div>
-        </div>
+  <div class="row">
+    <div class="col-md-12 text-center">
+      <h1>Without Redux</h1>
     </div>
+    <div class="col-md-6">
+      <div class="todolist not-done">
+        <h1>To-do list</h1>
+        <div class="input-group">
+          <input type="text" class="form-control add-todo" placeholder="Add todo" ng-model="$ctrl.todo">
+          <span class="input-group-btn">
+            <button class="btn btn-primary" type="button" ng-disabled="!$ctrl.todo" ng-click="$ctrl.addTodo()">Add</button>
+          </span>
+        </div>
+
+        <button class="btn btn-success marginTop10px" ng-click="$ctrl.markAllAsDone()">Mark all as done</button>
+        <button class="btn btn-danger marginTop10px" ng-click="$ctrl.deleteAllDone()">Delete all done</button>
+        <div class="checkbox pull-right">
+          <label>
+            <input type="checkbox" ng-model="$ctrl.showDone" />
+            Show done
+          </label>
+        </div>
+        <hr>
+
+        <ul id="sortable" class="list-unstyled">
+          <li class="ui-state-default" ng-repeat="todo in $ctrl.todos" ng-if="!(!$ctrl.showDone && todo.done)">
+            <div class="checkbox">
+              <label class="done-{{todo.done}}">
+                <input type="checkbox" ng-model="todo.done" />
+                {{todo.text}}
+              </label>
+              <button class="remove-item btn btn-danger btn-xs pull-right" ng-click="$ctrl.deleteTodo($index)">
+                <span class="glyphicon glyphicon-remove"></span>
+              </button>
+            </div>
+          </li>
+        </ul>
+        <div class="todo-footer">
+          <strong><span class="count-todos">{{$ctrl.getTotalTodos()}}</span></strong> total |
+          <strong><span class="count-todos">{{$ctrl.getTotalDoneTodos()}}</span></strong> done
+        </div>
+      </div>
+    </div>
+    <div class="col-md-6">
+      <div class="todolist">
+        <h1>Already done</h1>
+        <button class="btn btn-warning marginTop10px" ng-click="$ctrl.markAllAsDone(false)">Uncheck all</button>
+        <ul id="done-items" class="list-unstyled">
+          <li ng-repeat="doneTodo in $ctrl.getAllDone()">
+            {{doneTodo.text}}
+            <button class="remove-item btn btn-danger btn-xs pull-right" ng-click="$ctrl.toggleDone($index)">
+              <span class="glyphicon glyphicon-remove"></span>
+            </button>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </div>
 </div>
