--- step-3/client/app/components/noredux/noredux.js
@@ -1,11 +1,0 @@
-import angular from 'angular';
-import uiRouter from '@uirouter/angularjs';
-import noreduxComponent from './noredux.component';
-
-const noreduxModule = angular.module('noredux', [
-  uiRouter
-])
-
-.component('noredux', noreduxComponent);
-
-export default noreduxModule;
