+++ step-3/client/app/components/noredux/noredux.module.js
@@ -1,0 +1,8 @@
+import angular from 'angular';
+import uiRouter from '@uirouter/angularjs';
+import { NoReduxComponent } from './noredux.component';
+
+export const NoReduxModule = angular
+  .module('noredux', [uiRouter])
+  .component('noredux', NoReduxComponent)
+  .name;
