--- step-3/client/app/components/withredux/withredux.js
@@ -1,11 +1,0 @@
-import angular from 'angular';
-import uiRouter from '@uirouter/angularjs';
-import withReduxComponent from './withredux.component';
-
-const withReduxModule = angular.module('withredux', [
-  uiRouter
-])
-
-.component('withRedux', withReduxComponent);
-
-export default withReduxModule;
