+++ step-3/client/app/constants/todos.js
@@ -2,14 +2,14 @@
  * Enumerable to be used on the reducer
  */
 export const TODOS = {
-    ADD_TODO: 'ADD_TODO',
-    REMOVE_TODO: 'REMOVE_TODO',
-    MARK_ALL_AS_DONE: 'MARK_ALL_AS_DONE',
-    REMOVE_ALL_DONE: 'REMOVE_ALL_DONE',
-    TOGGLE_SHOW_DONE: 'TOGGLE_SHOW_DONE',
-    TOGGLE_DONE: 'TOGGLE_DONE',
-    CLEAN_ERROR: 'CLEAN_ERROR',
-    SET_LOADING: 'SET_LOADING',
-    SET_NOTIFICATION: 'SET_NOTIFICATION'
+  ADD_TODO: 'ADD_TODO',
+  REMOVE_TODO: 'REMOVE_TODO',
+  MARK_ALL_AS_DONE: 'MARK_ALL_AS_DONE',
+  REMOVE_ALL_DONE: 'REMOVE_ALL_DONE',
+  TOGGLE_SHOW_DONE: 'TOGGLE_SHOW_DONE',
+  TOGGLE_DONE: 'TOGGLE_DONE',
+  CLEAN_ERROR: 'CLEAN_ERROR',
+  SET_LOADING: 'SET_LOADING',
+  SET_NOTIFICATION: 'SET_NOTIFICATION'
 };
