+++ step-3/client/app/reducers/index.js
@@ -1,11 +1,6 @@
 import { combineReducers } from 'redux';
 import { TodosReducer } from './todos.reducer';
 
-
-
-
 export const RootReducer = combineReducers({
-    TodosState: TodosReducer
-    // ,
-    // testingState: TestingReducer
+  TodosState: TodosReducer
 });
