+++ step-3/client/app/selectors/base.js
@@ -1,4 +1,3 @@
-
 import * as _ from 'lodash';
 
 export const createCombinedSelector = (...args) => {
