+++ step-3/client/index.html
@@ -1,18 +1,18 @@
 <!doctype html>
 <html lang="en">
 
-<head>
-  <meta charset="utf-8">
-  <title>Angular / Redux TODO app</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="description" content="Redux app">
-  <base href="/">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-</head>
+  <head>
+    <meta charset="utf-8">
+    <title>Angular / Redux TODO app</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="description" content="Redux app">
+    <base href="/">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+  </head>
 
-<body ng-app="app" ng-strict-di ng-cloak>
-  <!-- Our app entry point, everything will be generated here-->
-  <div ui-view></div>
-</body>
+  <body ng-app="app" ng-strict-di ng-cloak>
+    <!-- Our app entry point, everything will be generated here-->
+    <div ui-view></div>
+  </body>
 
 </html>
